split out reconcile into its own file
